requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
frozenlist==1.6.0
idna==3.10
lxml==5.4.0
multidict==6.4.3
openai==0.28.1
propcache==0.3.1
python-dotenv==1.0.0
requests==2.32.3
soupsieve==2.7
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
yarl==1.20.0
